requirements.txt
torch
torchvision
pytorch-lightning
timm
numpy
Pillow
